Aperture: Fix lint error in string containing '%' sign
Explanation for issues of type "StringFormatCount": When a formatted
string takes arguments, it usually needs to reference the same arguments
in all translations (or all arguments if there are no translations.

There are cases where this is not the case, so this issue is a warning
rather than an error by default. However, this usually happens when a
language is not translated or updated correctly.

Change-Id: If6be5a021ea1f02de56725f7279dfa514bb1c2c4